Minor bug fix to fn_shop.sqf
Having come across an issue with selling handguns and launchers I identified that the transfer of the magazine to the holder was trying to use the Array _x rather than the string _x#0 which had been defined earlier as _class. This fix corrects that minor error and allows magazines from handguns and launchers to now be added to the unit when selling said weapons.